Introduce production deployment scripts (#34)
Add scripts that can be used for starting to serve SmartCanvas application on a production cPouta virtual machine.

The added documentation includes commands for generating 3 management scripts with the effect of:
1. Updating deployed source code
2. Starting application
3. Stopping application

These management scripts can be executed on a new host straight after creating it.

The behavior of `scripts/deploy_production/production_start.sh` can be affected to a degree by modifying `scripts/deploy_production/config.sh`. At least `VIRTUAL_MACHINE_DNS_NAME` will need to be set to match the virtual machine running the scripts. The default config in currently contains the domain name of a testing virtual machine.