Propagate environment variable used in endpoint obfuscation
The changes result in SecretsUsedInArgOrEnv warning being generated when
building backend image. This becomes an issue if the image gets
published to a registry. The endpoint path can then be discovered from
the image. To reduce the visibility of endpoint path, pass it to
container by using bind mounts.

Created #37 for fixing SecretsUsedInArgOrEnv warnings.