Add front endpoint obfuscation and more (#39)
* add Env variable check to be used for endpoint obfuscation
* Propagate environment variable used in endpoint obfuscation

Additionally:
* Support deploying from main branch
* update consent image to be transparent
* Updated frontend dependencies
* Finetune consent form acceptance, remove TEST gdpr button

Passing the endpoint token using ARG in Dockerfile result in
SecretsUsedInArgOrEnv warning being generated when building backend
image. This becomes an issue if the image gets published to a registry.
The endpoint path can then be discovered from the image. To reduce the
visibility of endpoint path, pass it to container by using bind mounts.

Created #37 for fixing SecretsUsedInArgOrEnv warnings.

---------

Co-authored-by: Marko Tiitto <Marko.Tiitto@student.oulu.fi>